Remove memory manager
