+++ pkg/hintrunner/hint_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func TestAllocSegment(t *testing.T) {
-	vm, _ := defaultVirtualMachine()
+	vm := defaultVirtualMachine()
 	vm.Context.Ap = 3
 	vm.Context.Fp = 0
 
@@ -41,7 +41,7 @@ func TestAllocSegment(t *testing.T) {
 }
 
 func TestTestLessThanFalse(t *testing.T) {
-	vm, _ := defaultVirtualMachine()
+	vm := defaultVirtualMachine()
 	vm.Context.Ap = 0
 	vm.Context.Fp = 0
 	writeTo(vm, VM.ExecutionSegment, 0, memory.MemoryValueFromInt(17))
@@ -69,7 +69,7 @@ func TestTestLessThanFalse(t *testing.T) {
 }
 
 func TestTestLessThanTrue(t *testing.T) {
-	vm, _ := defaultVirtualMachine()
+	vm := defaultVirtualMachine()
 	vm.Context.Ap = 0
 	vm.Context.Fp = 0
 	writeTo(vm, VM.ExecutionSegment, 0, memory.MemoryValueFromInt(23))
