+++ pkg/runners/zero/zero.go
@@ -14,10 +14,10 @@ import (
 
 type ZeroRunner struct {
 	// core components
-	program       *Program
-	vm            *vm.VirtualMachine
-	hintrunner    hintrunner.HintRunner
-	memoryManager *memory.MemoryManager
+	program    *Program
+	vm         *vm.VirtualMachine
+	hintrunner hintrunner.HintRunner
+	memory     *memory.Memory
 	// config
 	proofmode bool
 	maxsteps  uint64
@@ -27,22 +27,22 @@ type ZeroRunner struct {
 
 // Creates a new Runner of a Cairo Zero program
 func NewRunner(program *Program, proofmode bool, maxsteps uint64) (*ZeroRunner, error) {
-	memoryManager := memory.CreateMemoryManager()
-	_, err := memoryManager.Memory.AllocateSegment(program.Bytecode) // ProgramSegment
+	memory := memory.InitializeEmptyMemory()
+	_, err := memory.AllocateSegment(program.Bytecode) // ProgramSegment
 	if err != nil {
 		return nil, err
 	}
-	memoryManager.Memory.AllocateEmptySegment() // ExecutionSegment
+	memory.AllocateEmptySegment() // ExecutionSegment
 
 	// todo(rodro): given the program get the appropiate hints
 	hintrunner := hintrunner.NewHintRunner(make(map[uint64]hintrunner.Hinter))
 
 	return &ZeroRunner{
-		program:       program,
-		hintrunner:    hintrunner,
-		memoryManager: memoryManager,
-		proofmode:     proofmode,
-		maxsteps:      maxsteps,
+		program:    program,
+		hintrunner: hintrunner,
+		memory:     memory,
+		proofmode:  proofmode,
+		maxsteps:   maxsteps,
 	}, nil
 }
 
@@ -96,7 +96,7 @@ func (runner *ZeroRunner) InitializeMainEntrypoint() (memory.MemoryAddress, erro
 	}
 
 	returnFp := memory.MemoryValueFromSegmentAndOffset(
-		runner.memory().AllocateEmptySegment(),
+		runner.memory.AllocateEmptySegment(),
 		0,
 	)
 	return runner.InitializeEntrypoint("main", nil, &returnFp)
@@ -115,7 +115,7 @@ func (runner *ZeroRunner) InitializeEntrypoint(
 		stack = append(stack, memory.MemoryValueFromFieldElement(arguments[i]))
 	}
 	end := memory.MemoryAddress{
-		SegmentIndex: uint64(runner.memory().AllocateEmptySegment()),
+		SegmentIndex: uint64(runner.memory.AllocateEmptySegment()),
 		Offset:       0,
 	}
 	stack = append(stack, *returnFp, memory.MemoryValueFromMemoryAddress(&end))
@@ -140,7 +140,7 @@ func (runner *ZeroRunner) initializeVm(initialPC *memory.MemoryAddress, stack []
 		Pc: *initialPC,
 		Ap: offset + uint64(len(stack)),
 		Fp: offset + uint64(len(stack)),
-	}, runner.memoryManager.Memory, vm.VirtualMachineConfig{ProofMode: runner.proofmode})
+	}, runner.memory, vm.VirtualMachineConfig{ProofMode: runner.proofmode})
 	return err
 }
 
@@ -191,15 +191,11 @@ func (runner *ZeroRunner) BuildProof() ([]byte, []byte, error) {
 		return nil, nil, err
 	}
 
-	return EncodeTrace(relocatedTrace), EncodeMemory(runner.memoryManager.RelocateMemory()), nil
-}
-
-func (runner *ZeroRunner) memory() *memory.Memory {
-	return runner.memoryManager.Memory
+	return EncodeTrace(relocatedTrace), EncodeMemory(runner.vm.RelocateMemory()), nil
 }
 
 func (runner *ZeroRunner) segments() []*memory.Segment {
-	return runner.memoryManager.Memory.Segments
+	return runner.memory.Segments
 }
 
 func (runner *ZeroRunner) pc() memory.MemoryAddress {
