--- pkg/vm/memory/memory_manager_test.go
@@ -1,150 +1,0 @@
-package memory
-
-import (
-	"testing"
-
-	f "github.com/consensys/gnark-crypto/ecc/stark-curve/fp"
-	"github.com/stretchr/testify/require"
-)
-
-func TestMemoryRelocationWithFelt(t *testing.T) {
-	// segment 0: [2, -, -, 3]
-	// segment 3: [5, -, 7, -, 11, 13]
-	// relocated: [-, 2, -, -, 3, 5, -, 7, -, 11, 13]
-
-	manager := CreateMemoryManager()
-	updateMemoryWithValues(
-		manager.Memory,
-		[]memoryWrite{
-			// segment zero
-			{0, 0, uint64(2)},
-			{0, 3, uint64(3)},
-			// segment three
-			{3, 0, uint64(5)},
-			{3, 2, uint64(7)},
-			{3, 4, uint64(11)},
-			{3, 5, uint64(13)},
-		},
-	)
-
-	res := manager.RelocateMemory()
-
-	expected := []*f.Element{
-		nil,
-		// segment zero
-		new(f.Element).SetUint64(2),
-		nil,
-		nil,
-		new(f.Element).SetUint64(3),
-		// segment three
-		new(f.Element).SetUint64(5),
-		nil,
-		new(f.Element).SetUint64(7),
-		nil,
-		new(f.Element).SetUint64(11),
-		new(f.Element).SetUint64(13),
-	}
-
-	require.Equal(t, len(expected), len(res))
-	require.Equal(t, expected, res)
-}
-
-func TestMemoryRelocationWithAddress(t *testing.T) {
-	// segment 0: [-, 1, -, 1:5] (4)
-	// segment 1: [1, 4:3, 7, -, -, 13] (10)
-	// segment 2: [0:1] (11)
-	// segment 3: [2:0] (12)
-	// segment 4: [0:0, 1:1, 1:5, 15] (16)
-	// relocated: [
-	//      dummy:  -,
-	//      zero:   -,  1, -, 10,
-	//      one:    1, 16, 7,  -, -, 13,
-	//      two:    2,
-	//      three: 11,
-	//      four:   1,  6, 10, 15,
-	// ]
-
-	manager := CreateMemoryManager()
-	updateMemoryWithValues(
-		manager.Memory,
-		[]memoryWrite{
-			// segment zero
-			{0, 1, uint64(1)},
-			{0, 3, &MemoryAddress{1, 5}},
-			// segment one
-			{1, 0, uint64(1)},
-			{1, 1, &MemoryAddress{4, 3}},
-			{1, 2, uint64(7)},
-			{1, 5, uint64(13)},
-			// segment two
-			{2, 0, &MemoryAddress{0, 1}},
-			// segment three
-			{3, 0, &MemoryAddress{2, 0}},
-			// segment four
-			{4, 0, &MemoryAddress{0, 0}},
-			{4, 1, &MemoryAddress{1, 1}},
-			{4, 2, &MemoryAddress{1, 5}},
-			{4, 3, uint64(15)},
-		},
-	)
-
-	res := manager.RelocateMemory()
-
-	expected := []*f.Element{
-		nil,
-		// segment zero
-		nil,
-		new(f.Element).SetUint64(1),
-		nil,
-		new(f.Element).SetUint64(10),
-		// segment one
-		new(f.Element).SetUint64(1),
-		new(f.Element).SetUint64(16),
-		new(f.Element).SetUint64(7),
-		nil,
-		nil,
-		new(f.Element).SetUint64(13),
-		// segment two
-		new(f.Element).SetUint64(2),
-		// segment three
-		new(f.Element).SetUint64(11),
-		// segment 4
-		new(f.Element).SetUint64(1),
-		new(f.Element).SetUint64(6),
-		new(f.Element).SetUint64(10),
-		new(f.Element).SetUint64(15),
-	}
-
-	require.Equal(t, len(expected), len(res))
-	require.Equal(t, expected, res)
-}
-
-type memoryWrite struct {
-	SegmentIndex uint64
-	Offset       uint64
-	Value        any
-}
-
-func updateMemoryWithValues(memory *Memory, valuesToWrite []memoryWrite) {
-	var max_segment uint64 = 0
-	for _, toWrite := range valuesToWrite {
-		// wrap any inside a memory value
-		val, err := MemoryValueFromAny(toWrite.Value)
-		if err != nil {
-			panic(err)
-		}
-
-		// if the destination segment does not exist, create it
-		for toWrite.SegmentIndex >= max_segment {
-			max_segment += 1
-			memory.AllocateEmptySegment()
-		}
-
-		// write the memory val
-		err = memory.Write(toWrite.SegmentIndex, toWrite.Offset, &val)
-		if err != nil {
-			panic(err)
-		}
-
-	}
-}
